Don't explicitly enable multibuffer extension (use upstream default, which is "disabled" for now)
This extension is old, deprecated and has 2 events. Nvidia closed source driver
adds 5 events and makes us exceed the event limit (64). This causes random (and
probably important) extensions to be disabled.

This bug was triggered by the last rebuild, which uses the new dri2 proto, which
contains 1 more event.

Thanks to Anssi for debugging and finding the fix.

See bug #57889


SILENT: svn-revision: 518021